Highlighter.js converted to functional component
