+++ .changeset/two-llamas-film.md
@@ -1,0 +1,5 @@
+---
+"react-mentions": patch
+---
+
+Highlighter.js converted to functional component
