+++ src/Highlighter.js
@@ -1,4 +1,4 @@
-import React, { Component, Children } from 'react'
+import React, { Children, useState, useEffect } from 'react'
 import PropTypes from 'prop-types'
 import { defaultStyle } from './utils'
 
@@ -18,190 +18,147 @@ const _generateComponentKey = (usedKeys, id) => {
   return id + '_' + usedKeys[id]
 }
 
-class Highlighter extends Component {
-  static propTypes = {
-    selectionStart: PropTypes.number,
-    selectionEnd: PropTypes.number,
-    value: PropTypes.string.isRequired,
-    onCaretPositionChange: PropTypes.func.isRequired,
-    containerRef: PropTypes.oneOfType([
-      PropTypes.func,
-      PropTypes.shape({
-        current:
-          typeof Element === 'undefined'
-            ? PropTypes.any
-            : PropTypes.instanceOf(Element),
-      }),
-    ]),
-    children: PropTypes.oneOfType([
-      PropTypes.element,
-      PropTypes.arrayOf(PropTypes.element),
-    ]).isRequired,
-  }
+function Highlighter({
+  selectionStart,
+  selectionEnd,
+  value = '',
+  onCaretPositionChange,
+  containerRef,
+  children,
+  singleLine,
+  style,
+}) {
+  const [position, setPosition] = useState({ left: undefined, top: undefined })
+  const [caretElement, setCaretElement] = useState()
 
-  static defaultProps = {
-    value: '',
-  }
+  useEffect(() => {
+    notifyCaretPosition()
+  }, [caretElement])
 
-  constructor() {
-    super(...arguments)
-
-    this.state = { left: undefined, top: undefined }
-  }
-
-  componentDidMount() {
-    this.notifyCaretPosition()
-  }
-
-  componentDidUpdate() {
-    this.notifyCaretPosition()
-  }
-
-  notifyCaretPosition() {
-    if (!this.caretElement) {
+  const notifyCaretPosition = () => {
+    if (!caretElement) {
       return
     }
 
-    const { offsetLeft, offsetTop } = this.caretElement
+    const { offsetLeft, offsetTop } = caretElement
 
-    if (this.state.left === offsetLeft && this.state.top === offsetTop) {
+    if (position.left === offsetLeft && position.top === offsetTop) {
       return
     }
 
-    const newPosition = {
-      left: offsetLeft,
-      top: offsetTop,
-    }
-    this.setState(newPosition)
+    const newPosition = { left: offsetLeft, top: offsetTop }
+    setPosition(newPosition)
 
-    this.props.onCaretPositionChange(newPosition)
+    onCaretPositionChange(newPosition)
   }
 
-  render() {
-    const {
-      selectionStart,
-      selectionEnd,
+  const config = readConfigFromChildren(children)
+  let caretPositionInMarkup
+
+  if (selectionEnd === selectionStart) {
+    caretPositionInMarkup = mapPlainTextIndex(
       value,
-      style,
-      children,
-      containerRef,
-    } = this.props
-    const config = readConfigFromChildren(children)
-
-    // If there's a caret (i.e. no range selection), map the caret position into the marked up value
-    let caretPositionInMarkup
-    if (selectionStart === selectionEnd) {
-      caretPositionInMarkup = mapPlainTextIndex(
-        value,
-        config,
-        selectionStart,
-        'START'
-      )
-    }
-
-    const resultComponents = []
-    const componentKeys = {}
-
-    // start by appending directly to the resultComponents
-    let components = resultComponents
-    let substringComponentKey = 0
-
-    const textIteratee = (substr, index, indexInPlainText) => {
-      // check whether the caret element has to be inserted inside the current plain substring
-      if (
-        isNumber(caretPositionInMarkup) &&
-        caretPositionInMarkup >= index &&
-        caretPositionInMarkup <= index + substr.length
-      ) {
-        // if yes, split substr at the caret position and insert the caret component
-        const splitIndex = caretPositionInMarkup - index
-        components.push(
-          this.renderSubstring(
-            substr.substring(0, splitIndex),
-            substringComponentKey
-          )
-        )
-
-        // add all following substrings and mention components as children of the caret component
-        components = [
-          this.renderSubstring(
-            substr.substring(splitIndex),
-            substringComponentKey
-          ),
-        ]
-      } else {
-        // otherwise just push the plain text substring
-        components.push(this.renderSubstring(substr, substringComponentKey))
-      }
-
-      substringComponentKey++
-    }
-
-    const mentionIteratee = (
-      markup,
-      index,
-      indexInPlainText,
-      id,
-      display,
-      mentionChildIndex,
-      lastMentionEndIndex
-    ) => {
-      // generate a component key based on the id
-      const key = _generateComponentKey(componentKeys, id)
-      components.push(
-        this.getMentionComponentForMatch(id, display, mentionChildIndex, key)
-      )
-    }
-
-    iterateMentionsMarkup(value, config, mentionIteratee, textIteratee)
-
-    // append a span containing a space, to ensure the last text line has the correct height
-    components.push(' ')
-
-    if (components !== resultComponents) {
-      // if a caret component is to be rendered, add all components that followed as its children
-      resultComponents.push(this.renderHighlighterCaret(components))
-    }
-
-    return (
-      <div {...style} ref={containerRef}>
-        {resultComponents}
-      </div>
+      config,
+      selectionStart,
+      'START'
     )
   }
 
-  renderSubstring(string, key) {
-    // set substring span to hidden, so that Emojis are not shown double in Mobile Safari
+  const resultComponents = []
+  const componentKeys = {}
+  let components = resultComponents
+  let substringComponentKey = 0
+
+  const textIteratee = (substr, index, indexInPlainText) => {
+    if (
+      isNumber(caretPositionInMarkup) &&
+      caretPositionInMarkup >= index &&
+      caretPositionInMarkup <= index + substr.length
+    ) {
+      const splitIndex = caretPositionInMarkup - index
+      components.push(
+        renderSubstring(substr.substring(0, splitIndex), substringComponentKey)
+      )
+
+      components = [
+        renderSubstring(substr.substring(splitIndex), substringComponentKey),
+      ]
+    } else {
+      components.push(renderSubstring(substr, substringComponentKey))
+    }
+
+    substringComponentKey++
+  }
+
+  const mentionIteratee = (
+    markup,
+    index,
+    indexInPlainText,
+    id,
+    display,
+    mentionChildIndex,
+    lastMentionEndIndex
+  ) => {
+    const key = _generateComponentKey(componentKeys, id)
+    components.push(
+      getMentionComponentForMatch(id, display, mentionChildIndex, key)
+    )
+  }
+
+  const renderSubstring = (string, key) => {
     return (
-      <span {...this.props.style('substring')} key={key}>
+      <span {...style('substring')} key={key}>
         {string}
       </span>
     )
   }
 
-  // Returns a clone of the Mention child applicable for the specified type to be rendered inside the highlighter
-  getMentionComponentForMatch(id, display, mentionChildIndex, key) {
+  const getMentionComponentForMatch = (id, display, mentionChildIndex, key) => {
     const props = { id, display, key }
-    const child = Children.toArray(this.props.children)[mentionChildIndex]
+    const child = Children.toArray(children)[mentionChildIndex]
     return React.cloneElement(child, props)
   }
 
-  // Renders an component to be inserted in the highlighter at the current caret position
-  renderHighlighterCaret(children) {
+  const renderHighlighterCaret = (children) => {
     return (
-      <span
-        {...this.props.style('caret')}
-        ref={this.setCaretElement}
-        key="caret"
-      >
+      <span {...style('caret')} ref={setCaretElement} key="caret">
         {children}
       </span>
     )
   }
 
-  setCaretElement = (el) => {
-    this.caretElement = el
+  iterateMentionsMarkup(value, config, mentionIteratee, textIteratee)
+  components.push(' ')
+
+  if (components !== resultComponents) {
+    resultComponents.push(renderHighlighterCaret(components))
   }
+
+  return (
+    <div {...style} ref={containerRef}>
+      {resultComponents}
+    </div>
+  )
+}
+
+Highlighter.propTypes = {
+  selectionStart: PropTypes.number,
+  selectionEnd: PropTypes.number,
+  value: PropTypes.string.isRequired,
+  onCaretPositionChange: PropTypes.func.isRequired,
+  containerRef: PropTypes.oneOfType([
+    PropTypes.func,
+    PropTypes.shape({
+      current:
+        typeof Element === 'undefined'
+          ? PropTypes.any
+          : PropTypes.instanceOf(Element),
+    }),
+  ]),
+  children: PropTypes.oneOfType([
+    PropTypes.element,
+    PropTypes.arrayOf(PropTypes.element),
+  ]).isRequired,
 }
 
 const styled = defaultStyle(
